+++ src/agent/client.js
@@ -1,4 +1,4 @@
-import { _, Q, defer, lazy_require, log } from 'azk';
+import { _, Q, defer, lazy_require, log, publish } from 'azk';
 import { config, set_config } from 'azk';
 import { Agent } from 'azk/agent';
 import { AgentNotRunning } from 'azk/utils/errors';
@@ -127,19 +127,21 @@ var WebSocketClient = {
 };
 
 var Client = {
-  status() {
+  status(action_name) {
     var status_obj = {
       agent   : false,
       docker  : false,
       balancer: false,
     };
 
-    return defer((resolve, _reject, notify) => {
+    return defer((resolve) => {
       if (Agent.agentPid().running) {
-        notify({ type: "status", status: "running" });
+        publish("agent.client.status", { type: "status", status: "running" });
         status_obj.agent = true;
       } else {
-        notify({ type: "status", status: "not_running" });
+        if (action_name !== 'start') {
+          publish("agent.client.status", { type: "status", status: "not_running" });
+        }
       }
       resolve(status_obj);
     });
@@ -150,10 +152,10 @@ var Client = {
   },
 
   stop(opts) {
-    return defer((_resolve, _reject, notify) => {
-      notify({ type: "status", status: "stopping" });
+    return defer(() => {
+      publish("agent.client.stop.status", { type: "status", status: "stopping" });
       return Agent.stop(opts).then((result) => {
-        if (result) { notify({ type: "status", status: "stopped" }); }
+        if (result) { publish("agent.client.stop.status", { type: "status", status: "stopped" }); }
         return { agent: result };
       });
     });